+++ packages/sources/react/src/components/indicators/VtmnBadge/VtmnBadge.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import '@vtmn/css-badge/dist/index-with-vars.css';
 import { VtmnBadgeVariant } from './types';
-import { objectValuesToString } from '../../../utils/object';
 
 export interface VtmnBadgeProps extends React.ComponentPropsWithoutRef<'span'> {
   /**
@@ -28,7 +27,7 @@ export const VtmnBadge = ({
       className={`vtmn-badge vtmn-badge_variant--${variant}  ${
         className ? className : ''
       }`}
-      {...objectValuesToString(props)}
+      {...props}
     >
       {value && value > 99 ? '99+' : value || ''}
     </span>