+++ packages/sources/react/src/components/selection-controls/VtmnRadioButton/VtmnRadioButton.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import '@vtmn/css-radio-button/dist/index-with-vars.css';
-import { objectValuesToString } from '../../../utils/object';
 
 export interface VtmnRadioButtonProps
   extends React.ComponentPropsWithoutRef<'input'> {
@@ -47,7 +46,7 @@ export const VtmnRadioButton = ({
         id={identifier}
         checked={checked}
         disabled={disabled}
-        {...objectValuesToString(props)}
+        {...props}
       />
       <label htmlFor={identifier}>{labelText}</label>
     </div>