+++ packages/sources/react/src/components/structure/VtmnDivider/VtmnDivider.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import clsx from 'clsx';
 import '@vtmn/css-divider/dist/index-with-vars.css';
 import { VtmnDividerOrientation, VtmnDividerTextPosition } from './types';
-import { objectValuesToString } from '../../../utils/object';
 
 export interface VtmnDividerProps
   extends React.ComponentPropsWithoutRef<'div'> {
@@ -30,11 +29,11 @@ export const VtmnDivider = ({
     <div
       className={clsx(
         'vtmn-divider',
-        children && `vtmn-divider_text-position--${textPosition}`,
         `vtmn-divider_orientation--${orientation}`,
+        children && `vtmn-divider_text-position--${textPosition}`,
         className,
       )}
-      {...objectValuesToString(props)}
+      {...props}
       role="separator"
       aria-orientation={orientation}
     >