+++ packages/sources/react/src/guidelines/iconography/VtmnIcon/VtmnIcon.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { VitamixId } from '@vtmn/icons/dist/vitamix/font/vitamix';
 import { VtmnIconColor, VtmnIconSize, VtmnIconVariant } from './types';
-import { objectValuesToString } from '../../../utils/object';
 
 export interface VtmnIconProps extends React.HTMLAttributes<HTMLElement> {
   /**
@@ -75,7 +74,7 @@ export const VtmnIcon: React.FC<VtmnIconProps> = ({
         color: getIconColor(),
         ...style,
       }}
-      {...objectValuesToString(props)}
+      {...props}
     ></span>
   );
 };