+++ packages/sources/svelte/src/components/indicators/VtmnPrice/test/VtmnPriceWithSlots.svelte
@@ -1,0 +1,5 @@
+<script>
+  import VtmnPrice from '../VtmnPrice.svelte';
+</script>
+
+<VtmnPrice {...$$restProps}>100 €</VtmnPrice>