+++ packages/sources/svelte/src/components/indicators/VtmnTag/test/VtmnTagWithSlots.svelte
@@ -1,0 +1,5 @@
+<script>
+  import VtmnTag from '../VtmnTag.svelte';
+</script>
+
+<VtmnTag {...$$restProps}>Unit-test tag</VtmnTag>