+++ packages/sources/svelte/src/components/overlays/VtmnModal/VtmnModal.svelte
@@ -22,7 +22,7 @@
    */
   export let animationDisabled = false;
 
-  let className = '';
+  let className = undefined;
   /**
    * @type {string} Custom classes to apply to the component.
    */