+++ packages/sources/vue/src/components/index.ts
@@ -5,9 +5,11 @@ import VtmnSelect from './forms/VtmnSelect';
 import VtmnBadge from './indicators/VtmnBadge';
 import VtmnLoader from './indicators/VtmnLoader';
 import VtmnPrice from './indicators/VtmnPrice';
+import VtmnTag from './indicators/VtmnTag';
 import VtmnCheckbox from './selection-controls/VtmnCheckbox';
 import VtmnRadioButton from './selection-controls/VtmnRadioButton';
 import VtmnToggle from './selection-controls/VtmnToggle';
+import VtmnDivider from './structure/VtmnDivider';
 
 export {
   VtmnButton,
@@ -19,5 +21,7 @@ export {
   VtmnPrice,
   VtmnCheckbox,
   VtmnRadioButton,
+  VtmnTag,
   VtmnToggle,
+  VtmnDivider,
 };